refactor: remove unused agents
This PR removes old agents and replaces them with dummy aliases agent.
