+++ src/agents/abis/aliases.json
@@ -1,0 +1,1 @@
+["function setAlias()"]
