--- src/agents/abis/discussions.json
@@ -1,12 +1,0 @@
-[
-  "function addCategory(uint64 parent, string metadataURI)",
-  "function editCategory(uint64 category, string metadataURI)",
-  "function removeCategory(uint64 category)",
-  "function addTopic(string author, uint64 category, uint64 parent, string metadataURI)",
-  "function editTopic(uint64 topic, string metadataURI)",
-  "function removeTopic(uint64 topic)",
-  "function pinTopic(uint64 topic)",
-  "function unpinTopic(uint64 topic)",
-  "function vote(string voter, uint64 topic, uint64 choice)",
-  "function unvote(string voter, uint64 topic)"
-]
