--- src/agents/abis/profiles.json
@@ -1,4 +1,0 @@
-[
-  "function setProfile(string user, string metadataURI)",
-  "function setStatement(string user, string org, string metadataURI)"
-]
