--- src/agents/abis/votes.json
@@ -1,3 +1,0 @@
-[
-  "function vote(string space, string voter, uint64 proposalId, uint64 choice, uint64 chainId, string sig)"
-]
