+++ src/api/overrides.json
@@ -1,8 +1,1 @@
-{
-  "decimal_types": {
-    "BigDecimalVP": {
-      "p": 60,
-      "d": 0
-    }
-  }
-}
+{}
