+++ src/api/schema.gql
@@ -1,97 +1,3 @@
-scalar Id
-scalar Text
-scalar BigDecimalVP
-
-type Category {
-  id: Int!
-  name: String
-  about: Text
-  parent: Int!
-  topic_count: Int!
-  category_count: Int!
-  metadata_uri: String
-}
-
-type Topic {
-  id: Int!
-  title: String
-  content: Text
-  category: Category
-  author: User
-  pinned: Boolean
-  parent: Int
-  score: Float!
-  reply_count: Int!
-  vote_count: Int!
-  metadata_uri: String
-}
-
-type TopicVote {
+type Aliases {
   id: String!
-  voter: User
-  topic: Topic
-  choice: Int
-}
-
-type User {
-  id: String!
-  topic_count: Int!
-  topic_vote_count: Int!
-  profile: Profile
-  statements: [Statement] @derivedFrom(field: "user")
-}
-
-type Profile {
-  id: String!
-  user: User!
-  name: String
-  about: Text
-  twitter: String
-  discord: String
-  telegram: String
-  github: String
-  updated: Int
-}
-
-type Statement {
-  id: String!
-  user: User!
-  org: String!
-  statement: Text
-  status: String
-  updated: Int
-}
-
-type SXUser {
-  id: String!
-  proposal_count: Int!
-  vote_count: Int!
-  created: Int!
-}
-
-type SXSpace {
-  id: String!
-  vote_count: Int!
-}
-
-type SXProposal {
-  id: String!
-  scores_1: BigDecimalVP!
-  scores_2: BigDecimalVP!
-  scores_3: BigDecimalVP!
-  scores_total: BigDecimalVP!
-  vote_count: Int!
-}
-
-type Vote {
-  id: String!
-  voter: SXUser!
-  space: SXSpace!
-  proposal: Int!
-  choice: Int!
-  vp: BigDecimalVP!
-  created: Int!
-  tx: String!
-  chain_id: Int # not part of sx-api
-  sig: Text # not part of sx-api
 }
